app-plane/services/audit-logger/requirements.txt
# Audit Logger Service Dependencies

# RabbitMQ client
pika==1.3.2

# PostgreSQL client (direct connection instead of Supabase REST API)
psycopg2-binary==2.9.9

# Shared dependencies (logger configuration)
python-dotenv==1.0.0
